fix: scan in into uint32 for PG oid type

+++ pkg/pg/collectors.go
@@ -113,7 +113,7 @@ func PGStatStatements(pgPool *pgxpool.Pool) func(ctx context.Context, state *age
 		// Build map from rows
 		queryStats := make(map[string]utils.CachedPGStatStatement)
 		for rows.Next() {
-			var queryid, userid, dbid int64
+			var queryid, userid, dbid uint32 // OID type in PostgreSQL is uint32
 			var calls int
 			var totalExecTime float64
 			var rowCount int64
